refactor: wgpu_layer module
