+++ reverie-engine/src/texture.rs
@@ -4,7 +4,7 @@ use etagere::{size2, AtlasAllocator};
 use image::{GenericImage, RgbaImage};
 use slotmap::SlotMap;
 
-use crate::wgpu_layer::WgpuTexture;
+use crate::wgpu_layer::texture::WgpuTexture;
 
 #[derive(Debug)]
 /// テクスチャ
