+++ reverie-engine/src/wgpu_layer.rs
@@ -1,5 +1,5 @@
 //! wgpu をラップするモジュール
-use std::{borrow::Cow, num::NonZeroU32, ops::Range};
+use std::{borrow::Cow, num::NonZeroU32};
 
 use anyhow::Context;
 use nalgebra::{Matrix4, Scale3, Translation3};
@@ -10,37 +10,12 @@ use crate::{
     texture::{TextureId, TextureRegistry},
 };
 
-#[repr(C)]
-#[derive(Debug, Clone, Copy, bytemuck::Pod, bytemuck::Zeroable)]
-/// UV 座標を持つ頂点
-///
-/// * `position`: 頂点の位置
-/// * `uv`: UV 座標
-pub struct UvVertex {
-    pub position: [f32; 3],
-    pub uv: [f32; 2],
-}
+use texture::WgpuTexture;
+use vertex::UvVertex;
 
-impl UvVertex {
-    pub const fn desc() -> w::VertexBufferLayout<'static> {
-        w::VertexBufferLayout {
-            array_stride: size_of::<Self>() as w::BufferAddress,
-            step_mode: w::VertexStepMode::Vertex,
-            attributes: &[
-                w::VertexAttribute {
-                    offset: 0,
-                    shader_location: 0,
-                    format: w::VertexFormat::Float32x3,
-                },
-                w::VertexAttribute {
-                    offset: size_of::<[f32; 3]>() as w::BufferAddress,
-                    shader_location: 1,
-                    format: w::VertexFormat::Float32x2,
-                },
-            ],
-        }
-    }
-}
+pub(crate) mod buffer;
+pub(crate) mod texture;
+pub(crate) mod vertex;
 
 /// wgpu を使うためのリソースをまとめた構造体
 pub struct WgpuResource<'window> {
@@ -410,233 +385,4 @@ fn setup_render_pipeline(
     });
 
     Ok(render_pipeline)
-}
-
-#[derive(Debug)]
-/// 頂点バッファとインデックスバッファをまとめた構造体
-pub(crate) struct VertexIndexBuffer {
-    pub(crate) vertex_buffer: w::Buffer,
-    vertex_array: Vec<UvVertex>,
-    pub(crate) index_buffer: w::Buffer,
-    index_array: Vec<u16>,
-    pub(crate) index_buffer_range: Range<u32>,
-}
-
-impl VertexIndexBuffer {
-    pub fn new(
-        device: &w::Device,
-        max_vertices: usize,
-        max_indices: usize,
-        label: Option<&str>,
-    ) -> anyhow::Result<Self> {
-        let name_v = label.map(|label| format!("{label} [vertex part]"));
-        let name_i = label.map(|label| format!("{label} [index part]"));
-
-        let vertex_buffer = device.create_buffer(&w::BufferDescriptor {
-            label: name_v.as_deref(),
-            usage: w::BufferUsages::VERTEX | w::BufferUsages::COPY_DST,
-            size: (max_vertices * size_of::<UvVertex>()) as u64,
-            mapped_at_creation: false,
-        });
-        let index_buffer = device.create_buffer(&w::BufferDescriptor {
-            label: name_i.as_deref(),
-            usage: w::BufferUsages::INDEX | w::BufferUsages::COPY_DST,
-            size: (max_indices * size_of::<u16>()) as u64,
-            mapped_at_creation: false,
-        });
-
-        Ok(Self {
-            vertex_buffer,
-            vertex_array: Vec::with_capacity(max_vertices),
-            index_buffer,
-            index_array: Vec::with_capacity(max_indices),
-            index_buffer_range: 0..0,
-        })
-    }
-
-    pub fn start_update<'a>(&'a mut self, queue: &'a w::Queue) -> VertexIndexBufferUpdater<'a> {
-        VertexIndexBufferUpdater {
-            buffer: self,
-            queue,
-            vertex_update: 0..0,
-            index_update: 0..0,
-        }
-    }
-
-    fn send_to_gpu(
-        &self,
-        queue: &w::Queue,
-        vertex_update: Range<usize>,
-        index_update: Range<usize>,
-    ) {
-        if !vertex_update.is_empty() {
-            queue.write_buffer(
-                &self.vertex_buffer,
-                vertex_update.start as u64,
-                bytemuck::cast_slice(&self.vertex_array[vertex_update]),
-            );
-        }
-        if !index_update.is_empty() {
-            queue.write_buffer(
-                &self.index_buffer,
-                index_update.start as u64,
-                bytemuck::cast_slice::<u16, u8>(&self.index_array[index_update]),
-            );
-        }
-    }
-}
-
-#[derive(Debug)]
-/// [`VertexIndexBuffer`] の更新を行うための構造体。Drop されると GPU にデータを送信する
-pub(crate) struct VertexIndexBufferUpdater<'a> {
-    buffer: &'a mut VertexIndexBuffer,
-    queue: &'a w::Queue,
-    vertex_update: Range<usize>,
-    index_update: Range<usize>,
-}
-
-impl<'a> VertexIndexBufferUpdater<'a> {
-    pub fn vertex_mut(&mut self) -> &mut Vec<UvVertex> {
-        &mut self.buffer.vertex_array
-    }
-
-    pub fn index_mut(&mut self) -> &mut Vec<u16> {
-        &mut self.buffer.index_array
-    }
-
-    /// 更新した頂点バッファの範囲を設定する
-    pub fn set_vertex_update(&mut self, range: Range<usize>) {
-        self.vertex_update = range;
-    }
-
-    /// 更新したインデックスバッファの範囲を設定する
-    pub fn set_index_update(&mut self, range: Range<usize>) {
-        self.index_update = range;
-    }
-
-    /// インデックスバッファの描画範囲を設定する
-    pub fn set_render_range(&mut self, range: Range<u32>) {
-        self.buffer.index_buffer_range = range;
-    }
-}
-
-impl<'a> std::ops::Drop for VertexIndexBufferUpdater<'a> {
-    fn drop(&mut self) {
-        self.buffer.send_to_gpu(
-            self.queue,
-            self.vertex_update.clone(),
-            self.index_update.clone(),
-        )
-    }
-}
-
-#[derive(Debug)]
-/// wgpu で使うテクスチャ
-pub(crate) struct WgpuTexture {
-    pub(crate) texture: w::Texture,
-    pub(crate) view: w::TextureView,
-}
-
-impl WgpuTexture {
-    pub fn from_image(
-        device: &w::Device,
-        queue: &w::Queue,
-        image: &image::RgbaImage,
-        label: Option<&str>,
-    ) -> Self {
-        let (width, height) = image.dimensions();
-
-        let size = w::Extent3d {
-            width,
-            height,
-            depth_or_array_layers: 1,
-        };
-        let texture = device.create_texture(&wgpu::TextureDescriptor {
-            label,
-            size,
-            mip_level_count: 1,
-            sample_count: 1,
-            dimension: wgpu::TextureDimension::D2,
-            format: wgpu::TextureFormat::Rgba8UnormSrgb,
-            usage: wgpu::TextureUsages::TEXTURE_BINDING | wgpu::TextureUsages::COPY_DST,
-            view_formats: &[],
-        });
-
-        queue.write_texture(
-            texture.as_image_copy(),
-            image,
-            w::ImageDataLayout {
-                offset: 0,
-                bytes_per_row: Some(width * 4),
-                rows_per_image: Some(height),
-            },
-            size,
-        );
-
-        let view = texture.create_view(&Default::default());
-
-        Self { texture, view }
-    }
-
-    pub fn width(&self) -> u32 {
-        self.texture.width()
-    }
-
-    pub fn height(&self) -> u32 {
-        self.texture.height()
-    }
-
-    pub(crate) fn bind_group_layout(
-        device: &w::Device,
-        label: Option<&str>,
-        texture_binding: u32,
-        sampler_binding: u32,
-    ) -> w::BindGroupLayout {
-        device.create_bind_group_layout(&w::BindGroupLayoutDescriptor {
-            label,
-            entries: &[
-                w::BindGroupLayoutEntry {
-                    binding: texture_binding,
-                    visibility: w::ShaderStages::FRAGMENT,
-                    ty: w::BindingType::Texture {
-                        sample_type: w::TextureSampleType::Float { filterable: true },
-                        view_dimension: w::TextureViewDimension::D2,
-                        multisampled: false,
-                    },
-                    count: None,
-                },
-                w::BindGroupLayoutEntry {
-                    binding: sampler_binding,
-                    visibility: w::ShaderStages::FRAGMENT,
-                    ty: w::BindingType::Sampler(w::SamplerBindingType::Filtering),
-                    count: None,
-                },
-            ],
-        })
-    }
-
-    pub(crate) fn create_bind_group(
-        &self,
-        device: &w::Device,
-        label: Option<&str>,
-        layout: &w::BindGroupLayout,
-        sampler: &w::Sampler,
-        texture_binding: u32,
-        sampler_binding: u32,
-    ) -> w::BindGroup {
-        device.create_bind_group(&w::BindGroupDescriptor {
-            label,
-            layout,
-            entries: &[
-                w::BindGroupEntry {
-                    binding: texture_binding,
-                    resource: w::BindingResource::TextureView(&self.view),
-                },
-                w::BindGroupEntry {
-                    binding: sampler_binding,
-                    resource: w::BindingResource::Sampler(sampler),
-                },
-            ],
-        })
-    }
 }
